修复midi文件长度判断bug 修复 非MIDI事件(ff事件)0x54的类型解析不全导致的长度解析错误 修复多个synther时，有时发声无法被关闭的bug 增加了midi解析文件优化midi事件数量的能力 对editor 载入midi文件，增加了异步加载的能力，加载过程中，随时可以取消加载，这针对黑乐谱长时间加载，如果不想等待，可以取消 修改了c#示例 修改了获取时间的返回值为double 修复了load midi 在非阻塞模式下的一些线程互斥错误 修复调用stop()后，紧接着调用play()产生咔嚓音的bug 修复移除虚拟乐器channel找不到乐器的bug 增加midi解析时，超时不再解析返回功能

+++ Tau/src/core/Synth/Editor/MidiEditor.h
@@ -46,7 +46,7 @@ namespace tau
 
 
 		//获取播放时间
-		inline float GetPlaySec()
+		inline double GetPlaySec()
 		{
 			return curtPlaySec;
 		}
